separate registration into standalone module
