+++ backend/src/server.ts
@@ -1,16 +1,12 @@
-import Fastify, { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
+import Fastify from 'fastify';
 import {TypeBoxTypeProvider} from '@fastify/type-provider-typebox';
 import fs from 'fs';
 import path from 'path';
 import fastifyCookie from '@fastify/cookie';
 import fastifyJwt from '@fastify/jwt';
-import { AuthController } from './api/controllers/AuthController';
-import { UserController } from './api/controllers/UserController';
-import { MatchmakingController } from './api/controllers/MatchmakingController';
 import fastifyMultipart from '@fastify/multipart';
-import { registerSecure } from './utils/registerSecure';
-import { MatchController } from './api/controllers/MatchController';
 import { MatchmakingService } from './services/MatchmakingService';
+import { registerRoutes } from './utils/registerRoutes';
 
 // server init
 const fastify = Fastify(
@@ -24,10 +20,6 @@ const fastify = Fastify(
 	}
 ).withTypeProvider<TypeBoxTypeProvider>();
 
-//test route
-fastify.get('/ping', async (request, reply) => {
-  return { status: 'ok', message: 'Pong!' };
-});
 
 fastify.register(fastifyCookie);
 fastify.register(fastifyJwt, {
@@ -35,19 +27,7 @@ fastify.register(fastifyJwt, {
 });
 fastify.register(fastifyMultipart);
 
-const auth = new AuthController(fastify);
-auth.registerRoutes();
-
-registerSecure(fastify, '/user', UserController, 'registerRoutes');
-registerSecure(fastify, '/matchmaking', MatchmakingController, 'registerProtectedRoutes');
-registerSecure(fastify, '/match', MatchController, 'registerRoutes');
-
-// public: /matchmaking/process
-fastify.register(async (app) => {
-  const mmCtrl = new MatchmakingController(app);
-  mmCtrl.registerPublicRoutes();     // тут /process
-}, { prefix: '/matchmaking' });
-
+registerRoutes(fastify);
 
 //server start
 fastify.listen({ port: 3000, host: '0.0.0.0' }, (err, address) => {
