+++ backend/src/utils/registerRoutes.ts
@@ -1,0 +1,69 @@
+import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
+import { AuthController } from "../api/controllers/AuthController";
+import { UserController } from "../api/controllers/UserController";
+import { MatchmakingController } from "../api/controllers/MatchmakingController";
+import { MatchController } from "../api/controllers/MatchController";
+import { TournamentController } from "../api/controllers/TournamentController";
+import { TournamentMatchController } from "../api/controllers/TournamentMatchController";
+import { TournamentParticipantController } from "../api/controllers/TournamentParticipantController";
+
+type ControllerConstructor = new (app: FastifyInstance) => any;
+
+export function registerRoutes(fastify: FastifyInstance) {
+
+	//test route
+	fastify.get('/ping', async (request, reply) => {
+		return { status: 'ok', message: 'Pong!' };
+	});
+
+	const auth = new AuthController(fastify);
+	auth.registerRoutes();
+
+	registerSecure(fastify, '/user', UserController, 'registerRoutes');
+	registerSecure(fastify, '/matchmaking', MatchmakingController, 'registerProtectedRoutes');
+	registerSecure(fastify, '/match', MatchController, 'registerRoutes');
+	registerSecure(fastify, '/tournament', TournamentController, 'registerSecureRoutes');
+	registerSecure(fastify, '/tournament', TournamentMatchController, 'registerSecureRoutes');
+	registerSecure(fastify, '/tournament', TournamentParticipantController, 'registerSecureRoutes');
+
+	// public: /matchmaking/process
+	fastify.register(async (app) => {
+		const mmCtrl = new MatchmakingController(app);
+		mmCtrl.registerPublicRoutes();
+	}, { prefix: '/matchmaking' });
+
+	fastify.register(async (app) => {
+		const mmCtrl = new TournamentController(app);
+		mmCtrl.registerRoutes();
+	}, { prefix: '/tournament' });
+
+	fastify.register(async (app) => {
+		const mmCtrl = new TournamentMatchController(app);
+		mmCtrl.registerRoutes();
+	}, { prefix: '/tournament' });
+
+	fastify.register(async (app) => {
+		const mmCtrl = new TournamentParticipantController(app);
+		mmCtrl.registerRoutes();
+	}, { prefix: '/tournament' });
+}
+
+function registerSecure(
+	fastify: FastifyInstance,
+  	prefix: string,
+  	controllerCtor: ControllerConstructor,
+  	methodName: keyof InstanceType<typeof controllerCtor>) 
+{
+  fastify.register(async (app) => {
+    // общий preHandler
+	app.addHook('preHandler', async (request: FastifyRequest, reply: FastifyReply) => {
+		try {
+			await request.jwtVerify();
+		}
+		catch (err){
+			reply.status(401).send({message: 'Unauthorized'});
+		}
+	});
+    new controllerCtor(app)[methodName]();
+  }, { prefix });
+}
